# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wall St. Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Token {
  id: number;
  name: string;
  ticker: string;
  description: string;
  imageUrl?: string | null;
  bondingCurveTarget: number;
  solRaised: number;
  creatorWallet?: string | null;
  mintAddress?: string | null;
  migrated: boolean;
  createdAt: string;
}

export interface CreateTokenRequest {
  name: string;
  /** @maxLength 10 */
  ticker: string;
  description: string;
  imageUrl?: string | null;
  bondingCurveTarget: number;
  creatorWallet?: string | null;
  mintAddress?: string | null;
}

export interface UpdateProgressRequest {
  solRaised: number;
}

export interface TickerCheck {
  available: boolean;
  minutesRemaining?: number | null;
}

export interface TickerCooldownError {
  error: string;
  minutesRemaining: number;
}

export type CheckTickerParams = {
  ticker: string;
};