# lib/api-zod/src/generated/types/createTokenRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wall St. Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTokenRequest {
  name: string;
  /** @maxLength 10 */
  ticker: string;
  description: string;
  imageUrl?: string | null;
  bondingCurveTarget: number;
  creatorWallet?: string | null;
  mintAddress?: string | null;
}