# lib/api-zod/src/generated/types/tickerCooldownError.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wall St. Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TickerCooldownError {
  error: string;
  minutesRemaining: number;
}