# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wall St. Launchpad API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Token {
  id: number;
  name: string;
  ticker: string;
  description: string;
  imageUrl?: string | null;
  bondingCurveTarget: number;
  solRaised: number;
  creatorWallet?: string | null;
  mintAddress?: string | null;
  migrated: boolean;
  createdAt: Date;
}